:rocket: Add different environments
Add support to production/development/etc… environments.
To use it DJANGO_SETTINGS_MODULE env variable must be set.
