+++ fiocruz/settings/staging.py
@@ -1,0 +1,1 @@
+from .base import *
