+++ fiocruz/settings/testing.py
@@ -1,0 +1,1 @@
+from .base import *
