+++ fiocruz/urls.py
@@ -1,37 +1,23 @@
-"""fiocruz URL Configuration
+import os
 
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/1.11/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  url(r'^$', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  url(r'^$', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.conf.urls import url, include
-    2. Add a URL to urlpatterns:  url(r'^blog/', include('blog.urls'))
-"""
 from django.conf.urls import include, url
 from django.contrib import admin
+from django.conf import settings
 
-from .settings import DEBUG
 
 core_urlpatterns = [
     url(r'^admin/', admin.site.urls),
 ]
-
-third_party_urlpatterns = []
-if DEBUG:
-    import debug_toolbar
-    third_party_urlpatterns += [
-        url(r'^__debug__/', include(debug_toolbar.urls)),
-    ]
 
 my_urlpatterns = [
     url(r'^polls/', include('polls.urls')),
     url(r'^samples/', include('samples.urls')),
 ]
 
-urlpatterns = core_urlpatterns + third_party_urlpatterns + my_urlpatterns
+urlpatterns = core_urlpatterns + my_urlpatterns
+
+if 'debug_toolbar' in settings.INSTALLED_APPS:
+    import debug_toolbar
+    urlpatterns += [
+        url(r'^__debug__/', include(debug_toolbar.urls)),
+    ]
