+++ requirements/prod.txt
@@ -1,0 +1,16 @@
+-r base.txt
+appnope==0.1.0
+decorator==4.0.11
+django-debug-toolbar==1.7
+ipython==6.0.0
+ipython-genutils==0.2.0
+jedi==0.10.2
+pexpect==4.2.1
+pickleshare==0.7.4
+prompt-toolkit==1.0.14
+ptyprocess==0.5.1
+Pygments==2.2.0
+simplegeneric==0.8.1
+sqlparse==0.2.3
+traitlets==4.3.2
+wcwidth==0.1.7
